LayoutTests/platform/win/editing/selection/click-in-margins-inside-editable-div-expected.txt

Ensure cursor placement matches IE6/IE7/FF3 when clicking above/below lines in padding.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Selection is at node: [object Text] offset: 0
PASS Selection is at node: [object Text] offset: 1
PASS Selection is at node: [object Text] offset: 2
PASS Selection is at node: [object Text] offset: 0
PASS Selection is at node: [object Text] offset: 2
PASS Selection is at node: [object Text] offset: 0
PASS Selection is at node: [object Text] offset: 1
PASS Selection is at node: [object Text] offset: 2
PASS Selection is at node: [object Text] offset: 0
PASS Selection is at node: [object Text] offset: 1
PASS Selection is at node: [object Text] offset: 2
PASS Selection is at node: [object Text] offset: 0
PASS Selection is at node: [object Text] offset: 1
PASS Selection is at node: [object Text] offset: 2
PASS successfullyParsed is true

TEST COMPLETE
